* for 64bit.
git-svn-id: http://svn.sourceforge.jp/svnroot/modchxj/mod_chxj/branches/RELEASE_0_12_0@3200 1a406e8e-add9-4483-a2c8-d8cac5b7c224

+++ src/chxj_serf.c
@@ -168,7 +168,7 @@ s_handle_response(serf_request_t *UNUSED(request), serf_bucket_t *response, void
         rv = serf_bucket_read(hdrs, 2048, &data, &len);
         if (SERF_BUCKET_READ_ERROR(rv))
           return rv;
-        tmp_headers = apr_pstrcat(ctx->pool, tmp_headers, apr_psprintf(ctx->pool , "%.*s", len, data), NULL);
+        tmp_headers = apr_pstrcat(ctx->pool, tmp_headers, apr_psprintf(ctx->pool , "%.*s", (unsigned int)len, data), NULL);
         if (APR_STATUS_IS_EOF(rv)) {
           break;
         }
